DimmSalt/requirements.txt
Flask
requests
flask-cors
python-dotenv
gunicorn
